feat: add format supports audio
Adds a property to videoProperties that
format supports audio.
