+++ lib/Settings.tsx
@@ -1,8 +1,9 @@
 import React, { ChangeEventHandler, useCallback, useRef, useState } from 'react'
 import styled from 'styled-components'
 
-import { VapixParameters, Format } from './PlaybackArea'
+import { VapixParameters } from './PlaybackArea'
 import { Switch } from './components/Switch'
+import { Format } from './types'
 
 const SettingsMenu = styled.div`
   font-family: sans-serif;
