+++ lib/constants.ts
@@ -1,0 +1,8 @@
+import { Format } from './types'
+
+export const FORMAT_SUPPORTS_AUDIO: Record<Format, boolean> = {
+  RTP_H264: true,
+  RTP_JPEG: false,
+  MP4_H264: true,
+  JPEG: false,
+}
