+++ lib/index.ts
@@ -5,6 +5,7 @@ export * from './BasicPlayer'
 export * from './Container'
 export * from './PlaybackArea'
 export * from './Stats'
+export * from './types'
 export * from './utils'
 
 window.customElements.define('media-stream-player', MediaStreamPlayer)
