+++ lib/types.ts
@@ -1,0 +1,6 @@
+export enum Format {
+  'RTP_H264' = 'RTP_H264',
+  'RTP_JPEG' = 'RTP_JPEG',
+  'JPEG' = 'JPEG',
+  'MP4_H264' = 'MP4_H264',
+}
